refactored list entries code
