+++ vis/js/components/EntriesWrapper.js
