+++ vis/js/components/ListEntries.js
@@ -1,17 +1,19 @@
 import React from "react";
 import { connect } from "react-redux";
 
-import BasicListEntries from "./listentries/BasicListEntries";
-import ClassificationListEntries from "./listentries/ClassificationListEntries";
-import StandardListEntries from "./listentries/StandardListEntries";
-import EntriesWrapper from "./listentries/EntriesWrapper";
-
 import { useLocalizationContext } from "./LocalizationProvider";
 
 import { filterData } from "../utils/data";
 import { STREAMGRAPH_MODE } from "../reducers/chartType";
+import { mapDispatchToListEntriesProps } from "../utils/eventhandlers";
+
+import EntriesWrapper from "./EntriesWrapper";
+import BasicListEntry from "../templates/listentry/BasicListEntry";
+import ClassificationListEntry from "../templates/listentry/ClassificationListEntry";
+import StandardListEntry from "../templates/listentry/StandardListEntry";
 
 const ListEntries = ({
+  // data
   show,
   service,
   displayedData,
@@ -19,8 +21,31 @@ const ListEntries = ({
   searchSettings,
   filterSettings,
   isStreamgraph,
+  disableClicks,
+  // functions
+  handleZoomIn,
+  handleSelectPaper,
+  handleSelectPaperWithZoom,
 }) => {
   const localization = useLocalizationContext();
+
+  const handleTitleClick = (paper) => {
+    if (disableClicks) {
+      return;
+    }
+    if (!isStreamgraph) {
+      handleSelectPaperWithZoom(paper);
+    } else {
+      handleSelectPaper(paper);
+    }
+  };
+
+  const handleAreaClick = (paper) => {
+    if (disableClicks) {
+      return;
+    }
+    handleZoomIn(paper);
+  };
 
   if (!show) {
     return null;
@@ -29,7 +54,7 @@ const ListEntries = ({
   let showEmptyMessage = displayedData.length === 0;
   if (!isStreamgraph && showEmptyMessage) {
     // we have to perform knowledge map filtering here (it's different
-    // to the filtering in the List.js component - paper selection is 
+    // to the filtering in the List.js component - paper selection is
     // not taken into account)
     // if the list is empty, but there are visible papers in the zoomed
     // bubble, the message is not displayed
@@ -47,15 +72,20 @@ const ListEntries = ({
     );
   }
 
-  if (service === null || typeof service === "undefined") {
-    return <BasicListEntries displayedData={displayedData} />;
-  }
+  const ListEntryComponent = getListEntryComponent(service);
 
-  if (service.startsWith("linkedcat")) {
-    return <ClassificationListEntries displayedData={displayedData} />;
-  }
-
-  return <StandardListEntries displayedData={displayedData} />;
+  return (
+    <EntriesWrapper>
+      {displayedData.map((paper) => (
+        <ListEntryComponent
+          key={paper.safe_id}
+          paper={paper}
+          handleTitleClick={() => handleTitleClick(paper)}
+          handleAreaClick={() => handleAreaClick(paper)}
+        />
+      ))}
+    </EntriesWrapper>
+  );
 };
 
 const mapStateToProps = (state) => ({
@@ -75,6 +105,22 @@ const mapStateToProps = (state) => ({
     docIds: state.selectedBubble ? state.selectedBubble.docIds : null,
   },
   isStreamgraph: state.chartType === STREAMGRAPH_MODE,
+  disableClicks: state.list.disableClicks,
 });
 
-export default connect(mapStateToProps)(ListEntries);
+export default connect(
+  mapStateToProps,
+  mapDispatchToListEntriesProps
+)(ListEntries);
+
+const getListEntryComponent = (service) => {
+  if (service === null || typeof service === "undefined") {
+    return BasicListEntry;
+  }
+
+  if (service.startsWith("linkedcat")) {
+    return ClassificationListEntry;
+  }
+
+  return StandardListEntry;
+};
