+++ vis/js/templates/listentry/StandardListEntry.jsx
@@ -1,4 +1,19 @@
 import React from "react";
+import { connect } from "react-redux";
+
+import { useLocalizationContext } from "../../components/LocalizationProvider";
+import { STREAMGRAPH_MODE } from "../../reducers/chartType";
+import {
+  getPaperComments,
+  getPaperKeywords,
+  getPaperPDFClickHandler,
+  getPaperPreviewLink,
+  getPaperTags,
+  getPaperTextLink,
+} from "../../utils/data";
+import { mapDispatchToListEntriesProps } from "../../utils/eventhandlers";
+import { shorten } from "../../utils/string";
+
 import Abstract from "./Abstract";
 import AccessIcons from "./AccessIcons";
 import Area from "./Area";
@@ -21,25 +36,76 @@ import Title from "./Title";
  * @param {Object} props
  */
 const StandardListEntry = ({
-  id,
-  access,
-  tags,
-  title,
-  preview,
-  details,
-  link,
-  documentType,
-  abstract,
-  comments,
-  keywords,
-  metrics,
-  area,
-  citations,
+  paper,
+  handlePDFClick,
+  linkType,
+  showDocumentType,
+  abstractSize,
+  showKeywords,
+  showMetrics,
+  isContentBased,
   baseUnit,
+  isStreamgraph,
+  handleAreaMouseover,
+  handleAreaMouseout,
   handleTitleClick,
   handleAreaClick,
-  backlink,
+  showBacklink,
+  isInStreamBacklink,
+  handleBacklinkClick,
+  // deprecated
+  showPreviewImage,
 }) => {
+  const loc = useLocalizationContext();
+
+  const id = paper.safe_id;
+  const access = {
+    isOpenAccess: !!paper.oa,
+    isFreeAccess: !!paper.free_access,
+    isDataset: paper.resulttype === "dataset",
+  };
+  const tags = getPaperTags(paper);
+  const title = paper.title ? paper.title : loc.default_paper_title;
+  const preview = {
+    link: getPaperPreviewLink(paper),
+    onClickPDF: getPaperPDFClickHandler(paper, handlePDFClick),
+    showPreviewImage,
+  };
+  const details = {
+    authors: paper.authors_string ? paper.authors_string : loc.default_authors,
+    source: paper.published_in,
+    year: paper.year,
+  };
+  const link = getPaperTextLink(paper, linkType);
+  const documentType = showDocumentType ? paper.resulttype : null;
+  const abstract = abstractSize
+    ? shorten(paper.paper_abstract, abstractSize)
+    : paper.paper_abstract;
+  const comments = getPaperComments(paper);
+  const keywords = showKeywords ? getPaperKeywords(paper, loc) : null;
+  const metrics = showMetrics
+    ? {
+        tweets: paper.cited_by_tweeters_count,
+        readers: paper["readers.mendeley"],
+        citations: paper.citation_count,
+        baseUnit: !isContentBased ? baseUnit : null,
+      }
+    : null;
+  const area = !isStreamgraph
+    ? {
+        text: paper.area,
+        onMouseOver: () => handleAreaMouseover(paper),
+        onMouseOut: () => handleAreaMouseout(),
+      }
+    : null;
+  const backlink = {
+    show: showBacklink,
+    isInStream: isInStreamBacklink,
+    onClick: () => handleBacklinkClick(),
+  };
+  const citations =
+    !isContentBased && !!baseUnit && !showMetrics ? paper.num_readers : null;
+
   return (
     // html template starts here
     <ListEntry anchorId={id}>
@@ -97,4 +163,24 @@ const StandardListEntry = ({
   );
 };
 
-export default StandardListEntry;
+const mapStateToProps = (state) => ({
+  abstractSize: state.selectedPaper ? null : state.list.abstractSize,
+  linkType: state.list.linkType,
+  showDocumentType: state.list.showDocumentType,
+  showMetrics: state.list.showMetrics,
+  isContentBased: state.list.isContentBased,
+  baseUnit: state.list.baseUnit,
+  showPreviewImage: !!state.selectedPaper,
+  showKeywords:
+    state.list.showKeywords &&
+    (!!state.selectedPaper || !state.list.hideUnselectedKeywords),
+  isStreamgraph: state.chartType === STREAMGRAPH_MODE,
+  showBacklink: state.chartType === STREAMGRAPH_MODE && !!state.selectedPaper,
+  isInStreamBacklink: !!state.selectedBubble,
+  disableClicks: state.list.disableClicks,
+});
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToListEntriesProps
+)(StandardListEntry);
